App::ElasticSearch::Utilities::HTTPRequest added.
There are a few places where a standard HTTP::Request object, or even an
HTTP::Request::JSON objects will barf on the payloads that we need to
send to Elasticsearch.  A::ES::U::HTTPRequest subclasses HTTP::Request
and overrides the new() and content() methods.

   * new() - Adds the Accept: application/json header.
   * content() - Handle how perl datastructures are encoded in the body.
     * HASH: Hashes are JSON encoded
     * ARRAY: Array elements are encoded as json and then those are
       joined together with "\n" (think the /_bulk API)
     * Non-reference Scalars: passed as is, you know what you're doing.
     * Everything else: discarded, generates a warning

The '_cat' API passes JSON back when you tell it you can read that, so
updated the es_indices() function to handle the JSON resulting from the
_cat call using the new A::ES::U::HTTPRequest object.