Make ::Query objects able to execute
This was an oversight on my part.  The `::Query` object now supports
`->execute( $list_of_indexes )` directly.  This makes working with the
library easier as folks don't need to read docs on the `es_request()`
function to figure out how to construct a search.

Additionally, `scroll_id`'s are automatically extracted from the result
set.  Calls to `->scroll_results()` will allow scrolling without having
to understand the mechanics of the scroll.