Adding es-aggregate.pl script
`es-search.pl` has a lot of cruft. Creating `es-aggregate.pl` as a
proof-of-concept starting from a clean slate to see if I can make the
code more readable and more reusable.  The idea being to push this code
back into the library to make working with aggregations easier and
cleaner.