Connection handling fixes
* Removed duplicate logic for version detection
* Added version hack for _cat/shards

The code to discover the version of the server duplicated some logic to
prevent an inifinte loop reusing `es_request()`. When I broke code out
into `::Connection`, the inifinite loop was no longer possible. This
removes the special casing and for version discovery and uses the
`::Connection->request()` method.

The version discovery code now fails more immediately and with better
error messages.